Fixed window drag for aspect2d & cleaned up gui code
